requirements.txt
defusedxml==0.5.0
entrypoints==0.2.3
Flask==1.0.2
pyTelegramBotAPI==3.7.1
Jinja2==2.11.2
MarkupSafe==0.23
Werkzeug==1.0.1
itsdangerous==1.1.0
suds-jurko==0.6
gunicorn==19.9.0
whitenoise==4.1.3
Telethon==1.13.0